requirements.txt
torch>=1.8.0
numpy>=1.24.0
pandas>=1.3.0
anthropic>=0.18.0
ta>=0.10.0
mplfinance>=0.12.0
pillow>=9.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pytest>=7.0.0
pytest-cov>=4.0.0
PyQt5>=5.15.0; sys_platform == 'win32'
